update code for album
