+++ .gitignore
@@ -27,3 +27,4 @@
 /config/master.key
 
 /storage
+public/system/
