+++ app/assets/javascripts/application.js
@@ -18,52 +18,12 @@
 //= require_tree .
 //= require bootstrap
 //= require activestorage
+
 import * as ActiveStorage from "activestorage";
-import "direct_uploads.js"
+import "./direct_uploads.js"
+
 ActiveStorage.start();
 
-$(document).ready(function(){
-//nut like
-  $('img.btn-like').click(function(){
-    if($(this).attr('src')=="/assets/heart.png")
-    {
-      $(this).removeAttr("src");
-      $(this).attr("src","/assets/like.png");
-    }
-    else
-    {
-      $(this).removeAttr("src");
-      $(this).attr("src","/assets/heart.png");
-    }
-  });
-//nut follow
-  $('button#btnFollow').click(function(){
-    if($(this).text() == "Follow"){
-      $(this).html("<b>Following</b>");
-      $(this).css("background-color", "#ffc107");
-      $(this).css("border-color", "white");
-      $(this).css("color", "white");
-    }else{
-      $(this).html("<b>Follow</b>");
-      $(this).css("background-color", "white");
-      $(this).css("border-color", "#ffc107");
-      $(this).css("color", "#ffc107");
-    }
-  });
-});
 
-//nut change anh dai dien
-document.querySelector("html").classList.add('js');
-function showImage(){
-  if(this.files && this.files[0])
-  {
-    var obj = new FileReader();
-    obj.onload = function(data){
-      var image = document.getElementById("image");
-      image.src = data.target.result;
-      image.style.display = "block"
-    }
-    obj.readAsDataURL(this.files[0]);
-  }
-}
+
