+++ app/controllers/albums_controller.rb
@@ -80,6 +80,6 @@ class AlbumsController < ApplicationController
 
     # Never trust parameters from the scary internet, only allow the white list through.
     def album_params
-      params.require(:album).permit(:title, :description, :sharing_mode, :delete_image_attachment, images: [])
+      params.require(:album).permit(:title, :description, :sharing_mode, images: [])
     end
 end
