+++ app/views/albums/_form.html.erb
@@ -11,7 +11,7 @@
               <hr>
             </div>
           </div>
-          <%= simple_form_for(@album) do |f| %>
+          <%= simple_form_for @album, multipart: true, remote: true do |f| %>
             <%= f.error_notification %>
             <%= f.error_notification message: f.object.errors[:base].to_sentence if f.object.errors[:base].present? %>
             <div class="form-inputs">
@@ -33,7 +33,6 @@
                   </div>
                 </div>
               </div>
-
               <div class="row">
                 <div class="col images">
                   <% if @album.images.exists? %>
