+++ app/views/photos/_form.html.erb
@@ -43,23 +43,21 @@
                     <div class="col images">
                       <div class="upload-btn-wrapper float-left size-my-photos">
                         <% if !@photo.image.attached? %>
-                          <div class="btn-upload" style="display: block;">
+                          <div class="btn-upload size-my-photos" style="display: block;">
                             <button class="btn-upload"><img src="/assets/plus4.png"></button>
                             <%= f.file_field :image, class:"btn-upload", onchange:"showImagePhoto.call(this)" %>
                           </div>
                       <% else %>
-                          <div class="img float-left mt-2 mr-1">
                             <div class="img float-left mt-2 mr-1">
                               <div ng-repeat="file in imagefinaldata" class="img-wrp size-my-photos">
-                                <%= image_tag @photo.image.variant(resize: "300x300") if @photo.image.attached? %>
-                                  <img class="btn-close m-1" src="/assets/close.png" onclick="clickClose.call(this)" />
+                                <%= image_tag @photo.image.variant(resize: "300x300"), {class:"size-my-photos"} if @photo.image.attached? %>
+                                <img class="btn-close m-1" src="/assets/close.png" onclick="clickClose.call(this)" />
                               </div>
-                              <div class="btn-upload" style="display: none;">
+                              <div class="btn-upload size-my-photos" style="display: none;">
                               <button class="btn-upload"><img src="/assets/plus4.png"></button>
                                 <%= f.file_field :image, class:"btn-upload", onchange:"showImagePhoto.call(this)" %>
                               </div>
                             </div>
-                          </div>
                         <% end %>
                       </div>
                     </div>
